# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PetConnect API
  description: |
    API REST para PetConnect — plataforma de mascotas que conecta dueños,
    veterinarios y comunidad. Incluye feed social, ranking semanal, InstaPet
    (red social por mascota), mascotas perdidas y adopciones.
  version: 0.1.0
  contact:
    name: PetConnect Team
servers:
  - url: http://localhost:8001/api/v1
    description: Servidor local de desarrollo (Hono)
  - url: https://api.petconnect.app/api/v1
    description: Producción
security:
  - bearerAuth: []
paths:
...
  /pets/{pet_id}/instapet/posts:
    get:
      tags: [InstaPet]
      summary: Listar posts del perfil InstaPet de una mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Posts paginados del perfil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInstaPetPosts"
    post:
      tags: [InstaPet]
      summary: Publicar en el perfil InstaPet de una mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstaPetPostRequest"
      responses:
        "201":
          description: Post creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstaPetPost"
        "422":
          $ref: "#/components/responses/ValidationError"
  /pets/{pet_id}/instapet/posts/{post_id}:
    get:
      tags: [InstaPet]
      summary: Obtener detalle de un post InstaPet
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/InstaPetPostId"
      responses:
        "200":
          description: Detalle del post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstaPetPostDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [InstaPet]
      summary: Eliminar post InstaPet
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/InstaPetPostId"
      responses:
        "204":
          description: Post eliminado
  /pets/{pet_id}/followers:
    get:
      tags: [InstaPet]
      summary: Listar seguidores de una mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Seguidores paginados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFollowers"
  /pets/{pet_id}/follow:
    post:
      tags: [InstaPet]
      summary: Seguir a una mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
      responses:
        "201":
          description: Ahora sigues a esta mascota
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [InstaPet]
      summary: Dejar de seguir a una mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
      responses:
        "204":
          description: Dejaste de seguir a esta mascota
  /me/following:
    get:
      tags: [InstaPet]
      summary: Listar mascotas que sigo
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Mascotas que sigo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFollowing"
  /pets/{pet_id}/milestones:
    get:
      tags: [InstaPet]
      summary: Listar hitos de vida de una mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Hitos paginados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMilestones"
    post:
      tags: [InstaPet]
      summary: Registrar un hito de vida
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneRequest"
      responses:
        "201":
          description: Hito registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstaPetMilestone"
        "422":
          $ref: "#/components/responses/ValidationError"

  # =========================================================================
  # SERVICES (MANADA LIBRE)
  # =========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT de Supabase Auth

  # ------------------------------------------------------------------
  # Parameters (reusables)
  # ------------------------------------------------------------------
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Número de página
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Resultados por página
    PetId:
      name: pet_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    InstaPetPostId:
      name: post_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  # ------------------------------------------------------------------
  # Common Responses
  # ------------------------------------------------------------------
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Conflicto (registro duplicado)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ------------------------------------------------------------------
  # Schemas
  # ------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Recurso no encontrado"
      required: [detail]
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        pages:
          type: integer
          example: 3

    # --- Auth ---
...
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          format: uuid
        username:
          type: string
          example: "petlover"
        avatar_url:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Pet ---
    Pet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
          example: "Luna"
        species:
          type: string
          example: "Perro"
        breed:
          type: string
          nullable: true
          example: "Golden Retriever"
        age:
          type: integer
          nullable: true
          example: 24
        weight:
          type: number
          nullable: true
          example: 25.5
        photo_url:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    InstaPetPost:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pet_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        photo_url:
          type: string
          format: uri
          nullable: true
        video_url:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        likes_count:
          type: integer
          example: 32
        comments_count:
          type: integer
          example: 5
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/Profile"
    InstaPetPostDetail:
      allOf:
        - $ref: "#/components/schemas/InstaPetPost"
        - type: object
          properties:
            pet:
              $ref: "#/components/schemas/Pet"
    CreateInstaPetPostRequest:
      type: object
      properties:
        photo_url:
          type: string
          format: uri
        video_url:
          type: string
          format: uri
        description:
          type: string
    PaginatedInstaPetPosts:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/InstaPetPost"

    # --- InstaPet Follower ---
    InstaPetFollower:
      type: object
      properties:
        id:
          type: string
          format: uuid
        follower_id:
          type: string
          format: uuid
        pet_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        follower:
          $ref: "#/components/schemas/Profile"
    FollowingPet:
      type: object
      properties:
        pet_id:
          type: string
          format: uuid
        pet_name:
          type: string
          example: "Luna"
        pet_photo_url:
          type: string
          format: uri
          nullable: true
        species:
          type: string
        followed_at:
          type: string
          format: date-time
    PaginatedFollowers:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/InstaPetFollower"
    PaginatedFollowing:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/FollowingPet"

    # --- InstaPet Milestone ---
    InstaPetMilestone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pet_id:
          type: string
          format: uuid
        title:
          type: string
          example: "Primer baño"
        description:
          type: string
          nullable: true
        photo_url:
          type: string
          format: uri
          nullable: true
        milestone_date:
          type: string
          format: date
          example: "2026-05-01"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateMilestoneRequest:
      type: object
      properties:
        title:
          type: string
          example: "Primer baño"
        description:
          type: string
        photo_url:
          type: string
          format: uri
        milestone_date:
          type: string
          format: date
      required: [title, milestone_date]
    PaginatedMilestones:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/InstaPetMilestone"

    # --- Services (Manada Libre) ---
...